Move const.py

+++ models/cnn_jet_img_encoder.py
@@ -2,7 +2,7 @@ from functools import reduce
 from typing import List, Optional, Tuple, Union
 import torch
 from torch import nn
-from .const import DEFAULT_DEVICE, DEFAULT_DTYPE
+from ..const import DEFAULT_DEVICE, DEFAULT_DTYPE
 from .dcnn import DCNN
 
 class CNNJetImgEncoder(nn.Module):
@@ -77,7 +77,7 @@ class CNNJetImgEncoder(nn.Module):
         :type flatten_hidden_widths: Optional[List[int]], optional
         :param device: Model's device, defaults to gpu if available, otherwise cpu.
         :type device: Optional[torch.device], optional
-        :param dtype: Model's data type, defaults to `torch.float64`
+        :param dtype: Model's data type, defaults to `torch.float`
         :type dtype: Optional[torch.dtype], optional
         """        
         super().__init__()
